+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Bug Fixes
 
 * [ObjectServer] Using Android Network Security Configuration is necessary to install the custom root CA for tests (API >= 24) (#5970).
-
+* Fixes issue with the incremental build causing direct access to model without accessor to fail (#6056).
 
 ## 5.3.1 (2018-06-19)